Moved the different advert modules (common/bankruptcy) under the advert module folder.
