+++ apps/legal-gazette-api/src/app/app.module.ts
@@ -21,13 +21,13 @@ import { AdvertModule } from '../modules/advert/advert.module'
 import { ApplicationModule } from '../modules/applications/application.module'
 import { BankruptcyApplicationModel } from '../modules/applications/bankruptcy/bankruptcy-application.model'
 import { CommonApplicationModule } from '../modules/applications/common/common-application.module'
-import { BankruptcyAdvertModel } from '../modules/bankruptcy/advert/bankruptcy-advert.model'
-import { BankruptcyDivisionAdvertModel } from '../modules/bankruptcy/division-advert/bankruptcy-division-advert.model'
+import { BankruptcyAdvertModel } from '../modules/advert/bankruptcy/advert/bankruptcy-advert.model'
+import { BankruptcyDivisionAdvertModel } from '../modules/advert/bankruptcy/division-advert/bankruptcy-division-advert.model'
 import { BaseEntityModule } from '../modules/base-entity/base-entity.module'
 import { CaseModel } from '../modules/case/case.model'
 import { CaseModule } from '../modules/case/case.module'
 import { CategoryModel } from '../modules/category/category.model'
-import { CommonAdvertModel } from '../modules/common-advert/common-advert.model'
+import { CommonAdvertModel } from '../modules/advert/common-advert/common-advert.model'
 import { CommunicationChannelModel } from '../modules/communication-channel/communication-channel.model'
 import { CourtDistrictModel } from '../modules/court-district/court-district.model'
 import { SettlementModel } from '../modules/settlement/settlement.model'
