+++ apps/legal-gazette-api/src/modules/advert/bankruptcy/division-advert/bankruptcy-division-advert.model.ts
@@ -11,8 +11,8 @@ import { z } from 'zod'
 import { LegalGazetteModels } from '@dmr.is/legal-gazette/constants'
 import { BaseModel, BaseTable } from '@dmr.is/shared/models/base'
 
-import { AdvertModel } from '../../advert/advert.model'
-import { SettlementModel } from '../../settlement/settlement.model'
+import { AdvertModel } from '../../advert.model'
+import { SettlementModel } from '../../../settlement/settlement.model'
 import { BankruptcyDivisionAdvertDto } from './dto/bankruptcy-division-advert.dto'
 
 export const bankruptcyDivisionAdvertSchema = z.object({
