+++ apps/legal-gazette-api/src/modules/advert/common-advert/dto/common-advert.dto.ts
