+++ apps/legal-gazette-api/src/modules/advert/common-advert/dto/update-common-advert.dto.ts
