+++ apps/legal-gazette-api/src/modules/applications/bankruptcy/bankruptcy-application.controller.ts
@@ -25,8 +25,8 @@ import { Auth } from '@island.is/auth-nest-tools'
 
 import { mapIndexToVersion } from '../../../lib/utils'
 import { AdvertModel } from '../../advert/advert.model'
-import { bankruptcyAdvertSchema } from '../../bankruptcy/advert/bankruptcy-advert.model'
-import { bankruptcyDivisionAdvertSchema } from '../../bankruptcy/division-advert/bankruptcy-division-advert.model'
+import { bankruptcyAdvertSchema } from '../../advert/bankruptcy/advert/bankruptcy-advert.model'
+import { bankruptcyDivisionAdvertSchema } from '../../advert/bankruptcy/division-advert/bankruptcy-division-advert.model'
 import { CaseModel } from '../../case/case.model'
 import { CaseDto } from '../../case/dto/case.dto'
 import { CategoryDefaultIdEnum } from '../../category/category.model'
