+++ apps/legal-gazette-api/src/modules/applications/bankruptcy/bankruptcy-application.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common'
 import { SequelizeModule } from '@nestjs/sequelize'
 
 import { AdvertModel } from '../../advert/advert.model'
-import { BankruptcyAdvertModel } from '../../bankruptcy/advert/bankruptcy-advert.model'
-import { BankruptcyDivisionAdvertModel } from '../../bankruptcy/division-advert/bankruptcy-division-advert.model'
+import { BankruptcyAdvertModel } from '../../advert/bankruptcy/advert/bankruptcy-advert.model'
+import { BankruptcyDivisionAdvertModel } from '../../advert/bankruptcy/division-advert/bankruptcy-division-advert.model'
 import { CaseModel } from '../../case/case.model'
 import { SettlementModel } from '../../settlement/settlement.model'
 import { BankruptcyApplicationController } from './bankruptcy-application.controller'
