+++ apps/legal-gazette-api/src/modules/applications/common/dto/common-application.dto.ts
@@ -10,7 +10,7 @@ import {
 
 import { ApiProperty } from '@nestjs/swagger'
 
-import { CreateCommonAdvertDto } from '../../../common-advert/dto/create-common-advert.dto'
+import { CreateCommonAdvertDto } from '../../../advert/common-advert/dto/create-common-advert.dto'
 
 enum CommonApplicationEventsEnum {
   APPROVE = 'APPROVE',
