+++ apps/legal-gazette-api/src/modules/case/case.model.ts
@@ -25,8 +25,8 @@ import {
   DeceasedApplicationCreateAttributes,
   DeceasedApplicationModel,
 } from '../applications/deceased/deceased-application.model'
-import { CommonAdvertModel } from '../common-advert/common-advert.model'
-import { CreateCommonAdvertInternalDto } from '../common-advert/dto/create-common-advert.dto'
+import { CommonAdvertModel } from '../advert/common-advert/common-advert.model'
+import { CreateCommonAdvertInternalDto } from '../advert/common-advert/dto/create-common-advert.dto'
 import {
   CommunicationChannelCreateAttributes,
   CommunicationChannelModel,
