+++ apps/legal-gazette-api/src/modules/settlement/settlement.model.ts
@@ -4,8 +4,8 @@ import { z } from 'zod'
 import { LegalGazetteModels } from '@dmr.is/legal-gazette/constants'
 import { BaseModel, BaseTable } from '@dmr.is/shared/models/base'
 
-import { BankruptcyAdvertModel } from '../bankruptcy/advert/bankruptcy-advert.model'
-import { BankruptcyDivisionAdvertModel } from '../bankruptcy/division-advert/bankruptcy-division-advert.model'
+import { BankruptcyAdvertModel } from '../advert/bankruptcy/advert/bankruptcy-advert.model'
+import { BankruptcyDivisionAdvertModel } from '../advert/bankruptcy/division-advert/bankruptcy-division-advert.model'
 import { SettlementDto } from './dto/settlement.dto'
 
 type SettlementAttributes = {
